[LP31718G42-62][Code maintenance] Creation of new model Tests, after the changes to implement the first physics algorithm

+++ src/main/java/lapr/project/model/Vehicle.java
@@ -384,7 +384,7 @@ public class Vehicle implements Serializable {
      * @param type the type of road that will be used as search criteria
      * @return the max velocity for specific type
      */
-    public double getVelocityLimit(String type) {
+    public int getVelocityLimit(String type) {
         return velocityLimitList.getVelocityLimit(type);
     }
